Remove mypy env from tox envlist
Since type annotanions not even close to complete at this stage
of development, we do not run mypy with tox for now.

Such restrictive mypy settings as we use slow down prototyping severely,
we should consider relaxing some of the settings or even drop mypy
altogether until the project is more or less mature.